requirements.txt
torch
transformers
Pillow
pytesseract
opencv-python
ultralytics
numpy